% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca.h.R
\name{lca}
\alias{lca}
\title{Latent Class Analysis}
\usage{
lca(data, dep, group, alt = "notequal", varEq = TRUE)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{group}{.}

\item{alt}{.}

\item{varEq}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Latent Class Analysis
}
