% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlca.h.R
\name{mlca}
\alias{mlca}
\title{Multilevel LCA}
\usage{
mlca(
  data,
  vars,
  covs,
  group,
  nc = 2,
  nclust = 2,
  fit = TRUE,
  comp1 = FALSE,
  rel1 = FALSE,
  cla = FALSE,
  cross = FALSE,
  mpc = FALSE,
  cpc = FALSE,
  cn = FALSE,
  co = FALSE,
  gof = FALSE,
  ci = FALSE,
  item = FALSE,
  plot1 = FALSE,
  plot2 = FALSE,
  angle = 0,
  plot3 = FALSE,
  width = 500,
  height = 500,
  width1 = 500,
  height1 = 500,
  width2 = 500,
  height2 = 500
)
}
\arguments{
\item{data}{the data as a data frame}

\item{vars}{.}

\item{covs}{.}

\item{group}{.}

\item{nc}{.}

\item{nclust}{.}

\item{fit}{.}

\item{comp1}{.}

\item{rel1}{.}

\item{cla}{.}

\item{cross}{.}

\item{mpc}{.}

\item{cpc}{.}

\item{cn}{.}

\item{co}{.}

\item{gof}{.}

\item{ci}{.}

\item{item}{.}

\item{plot1}{.}

\item{plot2}{.}

\item{angle}{a number from 0 to 90 defining the angle of the x-axis labels,
where 0 degrees represents completely horizontal labels.}

\item{plot3}{.}

\item{width}{.}

\item{height}{.}

\item{width1}{.}

\item{height1}{.}

\item{width2}{.}

\item{height2}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$progressBarHTML} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$fit} \tab \tab \tab \tab \tab a table \cr
  \code{results$comp1} \tab \tab \tab \tab \tab a table \cr
  \code{results$rel1} \tab \tab \tab \tab \tab a table \cr
  \code{results$cn} \tab \tab \tab \tab \tab a table \cr
  \code{results$cla} \tab \tab \tab \tab \tab a table \cr
  \code{results$cross} \tab \tab \tab \tab \tab a table \cr
  \code{results$mpc} \tab \tab \tab \tab \tab a table \cr
  \code{results$cpc} \tab \tab \tab \tab \tab a table \cr
  \code{results$gof} \tab \tab \tab \tab \tab a table \cr
  \code{results$ci} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
  \code{results$item} \tab \tab \tab \tab \tab an array of tables \cr
  \code{results$text3} \tab \tab \tab \tab \tab a preformatted \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$fit$asDF}

\code{as.data.frame(results$fit)}
}
\description{
Multilevel LCA
}
