% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.h.R
\name{profile}
\alias{profile}
\title{Profile plot}
\usage{
profile(data, vars, group, mc = TRUE, angle = 0, plot1 = FALSE)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{group}{.}

\item{mc}{.}

\item{angle}{a number from 0 to 90 defining the angle of the x-axis labels,
where 0 degrees represents completely horizontal labels.}

\item{plot1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$mc} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$mc$asDF}

\code{as.data.frame(results$mc)}
}
\description{
Profile plot
}
