% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radar.h.R
\name{radar}
\alias{radar}
\title{Radar chart}
\usage{
radar(data, labels, vars, num = 4, plot = FALSE, plot1 = FALSE, num1 = "4,6,8")
}
\arguments{
\item{data}{The data as a data frame.}

\item{labels}{.}

\item{vars}{.}

\item{num}{.}

\item{plot}{.}

\item{plot1}{.}

\item{num1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Radar chart
}
