% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain.h.R
\name{rain}
\alias{rain}
\title{Raincloud plot}
\usage{
rain(data, vars, group, angle = 0, plot = FALSE, width = 500, height = 500)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{group}{.}

\item{angle}{a number from 0 to 90 defining the angle of the x-axis labels,
where 0 degrees represents completely horizontal labels.}

\item{plot}{.}

\item{width}{.}

\item{height}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Raincloud plot
}
