% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.h.R
\name{step}
\alias{step}
\title{LCA/LPA 3-step approach}
\usage{
step(
  data,
  factors = list(list(label = "L[2]", vars = list())),
  covs,
  regform = "L ~ Species",
  impu = "modal",
  method = "BCH",
  run,
  reg = FALSE,
  fit = FALSE,
  par = FALSE,
  regform1 = "L ~ Species",
  impu1 = "modal",
  method1 = "BCH",
  reg1 = FALSE,
  plot = FALSE,
  plot1 = FALSE,
  ref = 0,
  ref1 = 0
)
}
\arguments{
\item{data}{.}

\item{factors}{a list containing named lists that define the \code{label}
of the factor and the \code{vars} that belong to that factor}

\item{covs}{.}

\item{regform}{.}

\item{impu}{.}

\item{method}{.}

\item{run}{.}

\item{reg}{.}

\item{fit}{.}

\item{par}{.}

\item{regform1}{.}

\item{impu1}{.}

\item{method1}{.}

\item{reg1}{.}

\item{plot}{.}

\item{plot1}{.}

\item{ref}{.}

\item{ref1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$fit} \tab \tab \tab \tab \tab a table \cr
  \code{results$post} \tab \tab \tab \tab \tab an output \cr
  \code{results$reg} \tab \tab \tab \tab \tab a table \cr
  \code{results$member} \tab \tab \tab \tab \tab an output \cr
  \code{results$pi} \tab \tab \tab \tab \tab a table \cr
  \code{results$rho} \tab \tab \tab \tab \tab a table \cr
  \code{results$reg1} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$fit$asDF}

\code{as.data.frame(results$fit)}
}
\description{
LCA/LPA 3-step approach
}
