% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.h.R
\name{word}
\alias{word}
\title{Wordcloud analysis}
\usage{
word(
  data,
  words,
  freq,
  minf = 1,
  maxw = 100,
  min = 1,
  max = 5,
  rot = 0.35,
  plot = TRUE,
  width = 500,
  height = 500,
  plot1 = FALSE
)
}
\arguments{
\item{data}{.}

\item{words}{.}

\item{freq}{.}

\item{minf}{.}

\item{maxw}{.}

\item{min}{.}

\item{max}{.}

\item{rot}{.}

\item{plot}{.}

\item{width}{.}

\item{height}{.}

\item{plot1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Wordcloud analysis
}
