% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glca.h.R
\name{glca}
\alias{glca}
\title{Multiple Group LCA}
\usage{
glca(
  data,
  vars,
  group,
  covs,
  nc = 2,
  nb = 50,
  fit = TRUE,
  comp = FALSE,
  rel = FALSE,
  invari = "T",
  marginal = FALSE,
  preval = FALSE,
  post = TRUE,
  item = TRUE,
  co = TRUE,
  plot1 = FALSE
)
}
\arguments{
\item{data}{the data as a data frame}

\item{vars}{.}

\item{group}{.}

\item{covs}{.}

\item{nc}{.}

\item{nb}{.}

\item{fit}{.}

\item{comp}{.}

\item{rel}{.}

\item{invari}{.}

\item{marginal}{.}

\item{preval}{.}

\item{post}{.}

\item{item}{.}

\item{co}{.}

\item{plot1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$fit} \tab \tab \tab \tab \tab a table \cr
  \code{results$comp} \tab \tab \tab \tab \tab a table \cr
  \code{results$rel} \tab \tab \tab \tab \tab a table \cr
  \code{results$marginal} \tab \tab \tab \tab \tab a table \cr
  \code{results$preval} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$text3} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$text1} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$fit$asDF}

\code{as.data.frame(results$fit)}
}
\description{
Multiple Group LCA
}
