% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glca.h.R
\name{glca}
\alias{glca}
\title{Multiple Group LCA}
\usage{
glca(
  data,
  vars,
  covs,
  group,
  nc = 2,
  fit = TRUE,
  mia = TRUE,
  mir = TRUE,
  cia = FALSE,
  cir = FALSE,
  marginal = FALSE,
  preval = FALSE,
  post = FALSE,
  item = FALSE,
  gamma = FALSE,
  co = FALSE,
  plot1 = FALSE
)
}
\arguments{
\item{data}{the data as a data frame}

\item{vars}{.}

\item{covs}{.}

\item{group}{.}

\item{nc}{.}

\item{fit}{.}

\item{mia}{.}

\item{mir}{.}

\item{cia}{.}

\item{cir}{.}

\item{marginal}{.}

\item{preval}{.}

\item{post}{.}

\item{item}{.}

\item{gamma}{.}

\item{co}{.}

\item{plot1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$fit} \tab \tab \tab \tab \tab a table \cr
  \code{results$mia} \tab \tab \tab \tab \tab a table \cr
  \code{results$mir} \tab \tab \tab \tab \tab a table \cr
  \code{results$cia} \tab \tab \tab \tab \tab a table \cr
  \code{results$cir} \tab \tab \tab \tab \tab a table \cr
  \code{results$marginal} \tab \tab \tab \tab \tab a table \cr
  \code{results$preval} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$text1} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$text3} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$text4} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$fit$asDF}

\code{as.data.frame(results$fit)}
}
\description{
Multiple Group LCA
}
