% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca.h.R
\name{lca}
\alias{lca}
\title{Latent Class Analysis}
\usage{
lca(data, vars, cov, class = 2, cluster = 1, lca = TRUE, lcr = FALSE)
}
\arguments{
\item{data}{the data as a data frame}

\item{vars}{.}

\item{cov}{.}

\item{class}{.}

\item{cluster}{.}

\item{lca}{.}

\item{lcr}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$text1} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Latent Class Analysis
}
