% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca.h.R
\name{lca}
\alias{lca}
\title{Latent Class Analysis}
\usage{
lca(vars, group, class = 2, cluster = 1, summary = TRUE)
}
\arguments{
\item{vars}{.}

\item{group}{.}

\item{class}{.}

\item{cluster}{.}

\item{summary}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
}
}
\description{
Latent Class Analysis
}
