% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca.h.R
\name{lca}
\alias{lca}
\title{Latent Class Analysis}
\usage{
lca(
  data,
  vars,
  covs,
  nc = 2,
  nb = 50,
  fit = TRUE,
  comp = FALSE,
  rel = FALSE,
  cp = FALSE,
  item = FALSE,
  gamma = FALSE,
  coef = FALSE,
  plot1 = FALSE,
  plot2 = FALSE,
  angle = 0,
  plot3 = FALSE,
  width = 500,
  height = 500
)
}
\arguments{
\item{data}{the data as a data frame}

\item{vars}{.}

\item{covs}{.}

\item{nc}{.}

\item{nb}{.}

\item{fit}{.}

\item{comp}{.}

\item{rel}{.}

\item{cp}{.}

\item{item}{.}

\item{gamma}{.}

\item{coef}{.}

\item{plot1}{.}

\item{plot2}{.}

\item{angle}{a number from 0 to 90 defining the angle of the x-axis labels,
where 0 degrees represents completely horizontal labels.}

\item{plot3}{.}

\item{width}{.}

\item{height}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$fit} \tab \tab \tab \tab \tab a table \cr
  \code{results$comp} \tab \tab \tab \tab \tab a table \cr
  \code{results$rel} \tab \tab \tab \tab \tab a table \cr
  \code{results$cp} \tab \tab \tab \tab \tab a table \cr
  \code{results$coef} \tab \tab \tab \tab \tab a table \cr
  \code{results$item} \tab \tab \tab \tab \tab an array of tables \cr
  \code{results$post} \tab \tab \tab \tab \tab an output \cr
  \code{results$member} \tab \tab \tab \tab \tab an output \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$fit$asDF}

\code{as.data.frame(results$fit)}
}
\description{
Latent Class Analysis
}
