% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca.h.R
\name{lca}
\alias{lca}
\title{Latent Class Analysis}
\usage{
lca(
  data,
  vars,
  group,
  nc = 2,
  nb = 100,
  comp = TRUE,
  cp = FALSE,
  item = FALSE,
  plot1 = FALSE
)
}
\arguments{
\item{data}{the data as a data frame}

\item{vars}{.}

\item{group}{.}

\item{nc}{.}

\item{nb}{.}

\item{comp}{.}

\item{cp}{.}

\item{item}{.}

\item{plot1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$comp} \tab \tab \tab \tab \tab a table \cr
  \code{results$cp} \tab \tab \tab \tab \tab a table \cr
  \code{results$post} \tab \tab \tab \tab \tab an output \cr
  \code{results$text1} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$comp$asDF}

\code{as.data.frame(results$comp)}
}
\description{
Latent Class Analysis
}
