% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lta.h.R
\name{lta}
\alias{lta}
\title{Latent Transition Analysis}
\usage{
lta(
  data,
  factors = list(list(label = "L1[2]", vars = list())),
  covs,
  regform = "pf ~ SEX+RACE",
  impu = "modal",
  method = "BCH",
  cons = "L1,L2,L3",
  reg = FALSE,
  par2 = FALSE,
  par3 = FALSE,
  fit1 = FALSE,
  plot = FALSE,
  width = 500,
  height = 500,
  plot1 = FALSE,
  plot2 = FALSE,
  width1 = 500,
  height1 = 500,
  width2 = 500,
  height2 = 500,
  tau = FALSE,
  stayer = FALSE
)
}
\arguments{
\item{data}{.}

\item{factors}{a list containing named lists that define the \code{label}
of the factor and the \code{vars} that belong to that factor}

\item{covs}{.}

\item{regform}{.}

\item{impu}{.}

\item{method}{.}

\item{cons}{.}

\item{reg}{.}

\item{par2}{.}

\item{par3}{.}

\item{fit1}{.}

\item{plot}{.}

\item{width}{.}

\item{height}{.}

\item{plot1}{.}

\item{plot2}{.}

\item{width1}{.}

\item{height1}{.}

\item{width2}{.}

\item{height2}{.}

\item{tau}{.}

\item{stayer}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$fit1} \tab \tab \tab \tab \tab a table \cr
  \code{results$reg} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$text3} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$text4} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$tau} \tab \tab \tab \tab \tab a table \cr
  \code{results$stay} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$fit1$asDF}

\code{as.data.frame(results$fit1)}
}
\description{
Latent Transition Analysis
}
