% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lta.h.R
\name{lta}
\alias{lta}
\title{Latent Transition Analysis}
\usage{
lta(
  data,
  vars,
  covs,
  form1 =
    "lc(6) ~ ESMK_98+FSMK_98+DSMK_98+HSMK_98+EDRK_98+CDRK_98+WDRK_98+BDRK_98+EMRJ_98+CMRJ_98+OMRJ_98+SMRJ_98",
  form2 = "lc ~ SEX",
  method = "naive",
  fit = FALSE,
  est = FALSE,
  desc = TRUE,
  cp = FALSE,
  plot1 = FALSE,
  plot = FALSE,
  raw = "FALSE",
  width = 500,
  height = 500,
  width1 = 500,
  height1 = 500
)
}
\arguments{
\item{data}{.}

\item{vars}{.}

\item{covs}{.}

\item{form1}{.}

\item{form2}{.}

\item{method}{.}

\item{fit}{.}

\item{est}{.}

\item{desc}{.}

\item{cp}{.}

\item{plot1}{.}

\item{plot}{.}

\item{raw}{.}

\item{width}{.}

\item{height}{.}

\item{width1}{.}

\item{height1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text1} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$text3} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$text4} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Latent Transition Analysis
}
