% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lta.h.R
\name{lta}
\alias{lta}
\title{Latent Transition Analysis}
\usage{
lta(
  data,
  factors = list(list(label = "L1[2]", vars = list())),
  covs,
  form1 = "L1 ~ SEX+RACE",
  method = "ML",
  cons = "L1,L2, L3",
  nc = 2,
  par = FALSE,
  reg = FALSE,
  par1 = FALSE
)
}
\arguments{
\item{data}{.}

\item{factors}{a list containing named lists that define the \code{label}
of the factor and the \code{vars} that belong to that factor}

\item{covs}{.}

\item{form1}{.}

\item{method}{.}

\item{cons}{.}

\item{nc}{.}

\item{par}{.}

\item{reg}{.}

\item{par1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$post} \tab \tab \tab \tab \tab an output \cr
  \code{results$reg} \tab \tab \tab \tab \tab a table \cr
  \code{results$member} \tab \tab \tab \tab \tab an output \cr
  \code{results$text1} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$text3} \tab \tab \tab \tab \tab a preformatted \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$reg$asDF}

\code{as.data.frame(results$reg)}
}
\description{
Latent Transition Analysis
}
