% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlca.h.R
\name{mlca}
\alias{mlca}
\title{Multilevel LCA}
\usage{
mlca(
  data,
  vars,
  group,
  nc = 2,
  nclust = 2,
  nb = 100,
  fit = TRUE,
  comp = TRUE,
  rel = FALSE,
  margin = FALSE,
  preval = FALSE,
  post = FALSE,
  item = TRUE,
  plot1 = FALSE
)
}
\arguments{
\item{data}{the data as a data frame}

\item{vars}{.}

\item{group}{.}

\item{nc}{.}

\item{nclust}{.}

\item{nb}{.}

\item{fit}{.}

\item{comp}{.}

\item{rel}{.}

\item{margin}{.}

\item{preval}{.}

\item{post}{.}

\item{item}{.}

\item{plot1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$fit} \tab \tab \tab \tab \tab a table \cr
  \code{results$comp} \tab \tab \tab \tab \tab a table \cr
  \code{results$rel} \tab \tab \tab \tab \tab a table \cr
  \code{results$preval} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$text1} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$fit$asDF}

\code{as.data.frame(results$fit)}
}
\description{
Multilevel LCA
}
