% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlca.h.R
\name{mlca}
\alias{mlca}
\title{Multilevel LCA}
\usage{
mlca(
  data,
  vars,
  covs,
  group,
  nc = 2,
  nclust = 2,
  nb = 10,
  fit = TRUE,
  comp1 = FALSE,
  rel1 = FALSE,
  margin = FALSE,
  cla = FALSE,
  cross = FALSE,
  co = FALSE,
  gof = FALSE,
  ci = FALSE,
  item = FALSE,
  post = FALSE,
  gamma = FALSE,
  member = FALSE,
  plot1 = FALSE,
  plot2 = FALSE,
  angle = 0,
  plot3 = FALSE
)
}
\arguments{
\item{data}{the data as a data frame}

\item{vars}{.}

\item{covs}{.}

\item{group}{.}

\item{nc}{.}

\item{nclust}{.}

\item{nb}{.}

\item{fit}{.}

\item{comp1}{.}

\item{rel1}{.}

\item{margin}{.}

\item{cla}{.}

\item{cross}{.}

\item{co}{.}

\item{gof}{.}

\item{ci}{.}

\item{item}{.}

\item{post}{.}

\item{gamma}{.}

\item{member}{.}

\item{plot1}{.}

\item{plot2}{.}

\item{angle}{a number from 0 to 90 defining the angle of the x-axis labels,
where 0 degrees represents completely horizontal labels.}

\item{plot3}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$fit} \tab \tab \tab \tab \tab a table \cr
  \code{results$comp1} \tab \tab \tab \tab \tab a table \cr
  \code{results$rel1} \tab \tab \tab \tab \tab a table \cr
  \code{results$margin} \tab \tab \tab \tab \tab a table \cr
  \code{results$cla} \tab \tab \tab \tab \tab a table \cr
  \code{results$cross} \tab \tab \tab \tab \tab a table \cr
  \code{results$gof} \tab \tab \tab \tab \tab a table \cr
  \code{results$ci} \tab \tab \tab \tab \tab a table \cr
  \code{results$member} \tab \tab \tab \tab \tab a table \cr
  \code{results$item} \tab \tab \tab \tab \tab an array of tables \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
  \code{results$text3} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$text4} \tab \tab \tab \tab \tab a preformatted \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$fit$asDF}

\code{as.data.frame(results$fit)}
}
\description{
Multilevel LCA
}
